refactor post execute for window, fix tests
