+++ ibis/pandas/execution/selection.py
@@ -77,7 +77,7 @@ def compute_projection_scalar_expr(expr, parent, data, scope=None, **kwargs):
 
 @compute_projection.register(ir.ColumnExpr, ops.Selection, pd.DataFrame)
 def compute_projection_column_expr(
-    expr, parent, data, scope=None, timecontext=None, **kwargs
+    expr, parent, data, scope, timecontext, **kwargs
 ):
     result_name = getattr(expr, '_name', None)
     op = expr.op()
@@ -279,9 +279,7 @@ def physical_tables_node(node):
 
 
 @execute_node.register(ops.Selection, pd.DataFrame)
-def execute_selection_dataframe(
-    op, data, scope=None, timecontext=None, **kwargs
-):
+def execute_selection_dataframe(op, data, scope, timecontext, **kwargs):
     selections = op.selections
     predicates = op.predicates
     sort_keys = op.sort_keys
