+++ ibis/pandas/execution/tests/test_timecontext.py
@@ -1,4 +1,3 @@
-import numpy as np
 import pandas as pd
 import pandas.util.testing as tm
 import pytest
@@ -11,11 +10,11 @@ pytestmark = pytest.mark.pandas
 def test_execute_with_timecontext(time_table):
     expr = time_table
     # define a time context for time-series data
-    context = ("20170101", "20170103")
+    context = pd.Timestamp("20170101"), pd.Timestamp("20170103")
 
     # without time context, execute produces every row
     df_all = expr.execute()
-    assert len(df_all["time"]) == 4
+    assert len(df_all["time"]) == 8
 
     # with context set, execute produces only rows within context
     df_within_context = expr.execute(timecontext=context)
@@ -35,7 +34,7 @@ def test_context_adjustment_asof_join(
     expr = time_keyed_left.asof_join(
         time_keyed_right, 'time', by='key', tolerance=2 * ibis.interval(days=1)
     )[time_keyed_left, time_keyed_right.other_value]
-    context = (3, 4)
+    context = pd.Timestamp(3), pd.Timestamp(4)
     result = expr.execute(timecontext=context)
 
     # compare with asof_join of manually trimmed tables
@@ -43,7 +42,7 @@ def test_context_adjustment_asof_join(
         time_keyed_df1["time"] >= pd.to_datetime(context[0])
     ][time_keyed_df1["time"] < pd.to_datetime(context[1])]
     trimmed_df2 = time_keyed_df2[
-        time_keyed_df2["time"] >= pd.to_datetime(context[0] - 2)
+        time_keyed_df2["time"] >= pd.to_datetime(context[0] - pd.Timedelta(2))
     ][time_keyed_df2["time"] < pd.to_datetime(context[1])]
     expected = pd.merge_asof(
         trimmed_df1,
@@ -55,23 +54,18 @@ def test_context_adjustment_asof_join(
     tm.assert_frame_equal(result, expected)
 
 
-def test_context_adjustment_window():
-    time = pd.date_range('20180101', '20180110')
-    start = 2
-    data = np.arange(start, start + len(time))
-    df = pd.DataFrame({'value': data, 'time': time}, index=time)
-    client = ibis.pandas.connect({'df': df})
-    t = client.table('df')
+def test_context_adjustment_window(time_table, time_df3):
     # trim data manually
     expected = (
-        df.set_index('time').value.rolling('3d', closed='both').mean()
-    )[4:].reset_index(drop=True)
+        time_df3.set_index('time').value.rolling('3d', closed='both').mean()
+    )[3:].reset_index(drop=True)
 
-    context = ('20180105', '20180111')
+    context = pd.Timestamp("20170105"), pd.Timestamp("20170111")
+
     expected.index.name = None
     day = ibis.interval(days=1)
-    window = ibis.trailing_window(3 * day, order_by=t.time)
-    expr = t.value.mean().over(window)
+    window = ibis.trailing_window(3 * day, order_by=time_table.time)
+    expr = time_table.value.mean().over(window)
     # result should adjust time context accordingly
     result = expr.execute(timecontext=context)
     tm.assert_series_equal(result, expected)
